Build fix?

+++ src/main/java/net/BukkitPE/network/protocol/SetTitlePacket.java
@@ -33,7 +33,7 @@ class SetTitlePacket extends DataPacket{
         this.putVarInt(this.fadeOutTime);
     }
 
-    public int pid(){
+    public byte pid(){
         return NETWORK_ID;
     }
 }
